requirements.txt
torch==2.0.0
torchvision
facenet-pytorch==2.5.3
pyyaml==5.4.1
torchmetrics